Renamed ReadMergeTemplates to just MergeTemplates.
