+++ src/Mailman.Server/ClientApp/src/actions/MergeTemplates.js
