+++ src/Mailman.Server/ClientApp/src/actions/index.js
@@ -1,6 +1,6 @@
 export * from './CreateMergeTemplate'
 export * from './Loading'
 export * from './Login'
-export * from './ReadMergeTemplates'
+export * from './MergeTemplates'
 export * from './Signalr'
 export * from './User'
